+++ lib/bloc/geoform_event.dart
@@ -5,9 +5,9 @@ abstract class GeoformEvent extends Equatable {
 }
 
 class ManualChanged extends GeoformEvent {
-  const ManualChanged({this.manual});
+  const ManualChanged({required this.manual});
 
-  final bool? manual;
+  final bool manual;
 
   @override
   List<Object?> get props => [manual];
@@ -40,4 +40,42 @@ class AddRegion extends GeoformEvent {
   List<Object?> get props => [
         region,
       ];
+}
+
+class ChangeMapPosition extends GeoformEvent {
+  const ChangeMapPosition({
+    this.position,
+  });
+  final LatLng? position;
+
+  @override
+  List<Object?> get props => [position];
+}
+
+class ChangeMarkers extends GeoformEvent {
+  const ChangeMarkers({
+    required this.markers,
+  });
+  final List<FastMarker> markers;
+
+  @override
+  List<Object?> get props => [markers];
+}
+
+class AddAnimation extends GeoformEvent {
+  const AddAnimation({
+    required this.controller,
+  });
+  final AnimationController controller;
+
+  @override
+  List<Object?> get props => [controller];
+}
+
+class ChangeActivateAction extends GeoformEvent {
+  const ChangeActivateAction({required this.isActivated});
+  final bool isActivated;
+
+  @override
+  List<Object?> get props => [isActivated];
 }